장고기능확장

+++ TOY_PJT/장고기능확장/articles/views.py
@@ -1,9 +1,9 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import article
 from .forms import MakeArticle
 from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import login_required
-
+from django.views.decorators.http import require_POST
 
 # Create your views here.
 
@@ -51,12 +51,30 @@ def update(request, pk):
 
 def detail(request, pk):
     pick_data = article.objects.get(pk=pk)
-    context = {"pick_data": pick_data}
+    like_count = pick_data.like_user.all()
+    context = {
+        "pick_data": pick_data,
+        "like_count": like_count,
+    }
 
     return render(request, "articles/detail.html", context)
 
 
 @login_required
 def delete(request, pk):
-    article.objects.get(pk=pk).delete()
+    if request.user.is_authenicated:
+        article.objects.get(pk=pk).delete()
     return redirect("articles:index")
+
+
+# 좋아요기능
+@require_POST
+def like_user(request, pk):
+    if request.user.is_authenticated:
+        a = get_object_or_404(article, pk=pk)
+        if a.like_user.filter(pk=request.user.pk).exists():
+            a.like_user.remove(request.user)
+        else:
+            a.like_user.add(request.user)
+        return redirect("articles:index")
+    return redirect("accounts:login")
